Added half days and limitations

+++ src/Controller/EventController.php
@@ -3,8 +3,12 @@
 namespace App\Controller;
 
 use App\Entity\Event;
+use App\Entity\EventType;
+use App\Entity\Holiday;
 use App\Entity\Status;
+use App\Entity\WorkCalendar;
 use App\Form\EventFormType;
+use DateTime;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -13,6 +17,7 @@ use Symfony\Component\Mailer\MailerInterface;
 use Symfony\Component\Mime\Email;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Component\Translation\TranslatableMessage;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * @Route("/{_locale}/event")
@@ -21,10 +26,12 @@ use Symfony\Component\Translation\TranslatableMessage;
 class EventController extends AbstractController
 {
     private $mailer = null;
+    private $translator = null;
 
-    public function __construct(MailerInterface $mailer)
+    public function __construct(MailerInterface $mailer, TranslatorInterface $translator)
     {
         $this->mailer = $mailer;
+        $this->translator = $translator;
     }
 
     /**
@@ -98,7 +105,13 @@ class EventController extends AbstractController
      */
     public function delete(Event $event = null): Response
     {
-        if (null !== $event) {
+        $days = $this->getParameter('days');
+        $interval = new \DateInterval("P${days}D");
+        $interval->invert = 1;
+        $deadline = (new \DateTime())->add($interval);
+        $deadlineStr = $deadline->format('Y-m-d 23:59:59');
+        $deadline = new \DateTime($deadlineStr);
+        if (null !== $event && $event->getStartDate() > $deadline) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($event);
             $em->flush();
@@ -112,9 +125,24 @@ class EventController extends AbstractController
                 $this->sendEmail($boss->getEmail(), $subject, $html, false);
             }
             return new Response(null, 204);
+        } else {
+            $message = $this->translator->trans('message.canNotDeletePastDay', [
+                'deadline' => $deadline->format('Y-m-d'),
+            ], 'messages');
+            //            $this->addFlash('error', $message);
+            $response = new Response($message, 422);
+            return $response;
         }
         $this->addFlash('error', 'message.eventNotFound');
         return $this->redirectToRoute('calendar');
+    }
+
+    /**
+     * @Route("/new", name="event_new", methods={"GET","POST"})
+     */
+    public function new(Request $request): Response
+    {
+        return $this->save($request);
     }
 
     /**
@@ -124,35 +152,42 @@ class EventController extends AbstractController
     {
         $event = new Event();
         $form = $this->createForm(EventFormType::class, $event, [
-            'days' => $this->getParameter('days')
+            'days' => $this->getParameter('days'),
+            'locale' => $request->getLocale(),
         ]);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
+            /** @var Event $event */
             $event = $form->getData();
             /** @var User $user */
             $user = $this->getUser();
-            $boss = $user->getBoss();
             if (null === $event->getId()) {
-                $myEvents = $this->getDoctrine()->getRepository(Event::class)->findUserEventsBeetweenDates($user, new \DateTime($event->getStartDate()->format('Y') . '-01-01'));
-                if ($this->checkOverlap($myEvents, $event)) {
-                    return $this->renderOverlapError($event, $form, $request->isXmlHttpRequest());
+                if ($event->getStartDate() > $event->getEndDate()) {
+                    $this->addFlash('error', 'message.startDateGreaterThanEndDate');
+                    $valid = false;
                 }
-                $event->setStatus($this->getDoctrine()->getRepository(Status::class)->find(Status::RESERVED));
-                $event->setUser($this->getUser());
-                return $this->renderSuccess($event, $boss, $request->isXmlHttpRequest());
+                if ($event->getUsePreviousYearDays()) {
+                    $year = intval($event->getStartDate()->format('Y')) - 1;
+                } else {
+                    $year = intval($event->getStartDate()->format('Y'));
+                }
+                /** @var WorkCalendar $workCalendar */
+                $workCalendar = $this->getDoctrine()->getRepository(WorkCalendar::class)->findOneBy(['year' => $year]);
+                // If any limitations exceeded it return the error directly
+                $valid = $this->checkDoesNotExcessLimitations($event, $workCalendar);
+                if ($valid) {
+                    $event->setStatus($this->getDoctrine()->getRepository(Status::class)->find(Status::RESERVED));
+                    $event->setUser($this->getUser());
+                    $event->setAskedAt(new \DateTime());
+                    $boss = $user->getBoss();
+                    return $this->renderSuccess($event, $boss, $request->isXmlHttpRequest());
+                } else {
+                    return $this->renderError($form, $request->isXmlHttpRequest());
+                }
             }
-            // For now is not necesary because can't edit an event
-            // else {
-            //     $bdEvent = $entityManager->getRepository(Event::class)->find($event->getId());
-            //     $bdEvent->fill($event);
-            //     $bdEvent->setUser($user);
-            //     return $this->renderSuccess($bdEvent, $boss, $request->isXmlHttpRequest());
-            // }
-            return $this->redirectToRoute('calendar');
         }
 
         $template = $request->isXmlHttpRequest() ? '_form.html.twig' : 'new.html.twig';
-
         return $this->render('event/' . $template, [
             'event' => $form->getData(),
             'form' => $form->createView(),
@@ -160,6 +195,64 @@ class EventController extends AbstractController
             null,
             $form->isSubmitted() && !$form->isValid() ? 422 : 200,
         ));
+    }
+
+    /**
+     * Returns valid if all limitations are passed. False in other case
+     */
+    private function checkDoesNotExcessLimitations(Event $event, WorkCalendar $workCalendar)
+    {
+        /* Check overlap with my own events */
+        $user = $this->getUser();
+        $myEvents = $this->getDoctrine()->getRepository(Event::class)->findUserEventsBeetweenDates($user, new \DateTime($event->getStartDate()->format('Y') . '-01-01'));
+        if ($this->checkOverlap($myEvents, $event)) {
+            $this->addFlash('error', new TranslatableMessage('message.overlapingDates', [
+                'startDate' => $event->getStartDate()->format('Y/m/d'),
+                'endDate' => $event->getEndDate()->format('Y/m/d')
+            ], 'messages'));
+            return false;
+        }
+        if ($event->getHalfDay() && $event->getStartDate() !== $event->getEndDate()) {
+            $this->addFlash('error', 'message.partitionableDaysOneByOne');
+            return false;
+        }
+        if ($event->getHalfDay() && ($event->getHours() < 2 || $event->getHours() > $workCalendar->getWorkingHours() / 2)) {
+            $this->addFlash('error', $this->translator->trans('message.partitionableHoursMinAndMax', [
+                'min' => "2",
+                'max' => number_format($workCalendar->getWorkingHours() / 2, 2),
+            ]));
+            return false;
+        }
+        $year = intval($event->getStartDate()->format('Y'));
+        if (new \DateTime() < new \DateTime("${year}-01-01") && ($event->getUsePreviousYearDays() === null || $event->getUsePreviousYearDays() === false)) {
+            $this->addFlash('error', $this->translator->trans('message.canNotAskBeforeDate', [
+                'year' => $year,
+                'startDate' => "${year}-01-01",
+            ]));
+            return false;
+        }
+        if ($event->getStartDate() > $workCalendar->getDeadlineNextYear()) {
+            $this->addFlash('error', $this->translator->trans('message.deadLineNextYearExceeded', [
+                'deadline' => $workCalendar->getDeadlineNextYear()->format('Y-m-d'),
+            ]));
+            return false;
+        }
+        if ($event->getType()->getId() !== EventType::PARTICULAR_BUSSINESS_LEAVE && $event->getHalfDay()) {
+            $this->addFlash('error', $this->translator->trans('message.partitionableDaysType', [
+                'hours' => $workCalendar->getPartitionableHours(),
+                'year' => $event->getStartDate()->format('Y'),
+            ]));
+            return false;
+        }
+        if ($event->getType()->getId() === EventType::PARTICULAR_BUSSINESS_LEAVE && $event->getHalfDay()) {
+            if (!$this->checkDoesNotExcessMaximumPartionableHours($event, $event->getStartDate()->format('Y'), $workCalendar)) {
+                return false;
+            }
+        }
+        if (!$this->checkDoesNotExcessMaximumDaysForType($event, $workCalendar)) {
+            return false;
+        }
+        return true;
     }
 
     private function sendEmail($to, $subject, $html, bool $sendToHHRR)
@@ -190,20 +283,6 @@ class EventController extends AbstractController
         return false;
     }
 
-    private function renderOverlapError($event, $form, $ajax = false)
-    {
-        $this->addFlash('error', new TranslatableMessage('message.overlapingDates', [
-            'startDate' => $event->getStartDate()->format('Y/m/d'),
-            'endDate' => $event->getEndDate()->format('Y/m/d')
-        ], 'messages'));
-        $template = $ajax ? '_form.html.twig' : 'new.html.twig';
-
-        return $this->render('event/' . $template, [
-            'event' => $form->getData(),
-            'form' => $form->createView(),
-        ], new Response(null, 422));
-    }
-
     private function renderSuccess($event, $boss, $ajax = false)
     {
         $entityManager = $this->getDoctrine()->getManager();
@@ -221,6 +300,122 @@ class EventController extends AbstractController
         }
         if ($ajax) {
             return new Response(null, 204);
+        } else {
+            return $this->redirectToRoute('myCalendar');
         }
+    }
+
+
+    private function calculateTotalWorkingDays(array $events, $workCalendar)
+    {
+        $totalWorkingDays = 0;
+        foreach ($events as $event) {
+            $workingDays = $this->calculateWorkingDays($event, $workCalendar);
+            $totalWorkingDays += $workingDays;
+        }
+
+        return $totalWorkingDays;
+    }
+
+    private function calculateWorkingDays(Event $event, WorkCalendar $workCalendar)
+    {
+        if (!$event->getHalfDay()) {
+            $holidaysBetween = count($this->getDoctrine()->getRepository(Holiday::class)->findHolidaysBetween($event->getStartDate(), $event->getEndDate()));
+            $workingDays = $event->getDays();
+            // Subtract two weekend days for every week in between
+            $weeks = floor($workingDays / 7);
+            //        dump($workingDays, $holidaysBetween, $weeks);
+            $workingDays -= $weeks * 2;
+            // Handle special cases
+            $startDay = intVal(date('w', strtotime(($event->getStartDate())->format('Y-m-d'))));
+            $endDay = intVal(date('w', strtotime(($event->getEndDate())->format('Y-m-d'))));
+            // Remove weekend not previously removed.   
+            if ($startDay - $endDay > 1) {
+                $workingDays -= 2;
+            }
+            // Remove start day if span starts on Sunday but ends before Saturday
+            if ($startDay == 7 && $endDay != 6) {
+                $workingDays--;
+            }
+            // Remove end day if span ends on Saturday but starts after Sunday
+            if ($endDay == 7 && $startDay != 1) {
+                $workingDays--;
+            }
+            $workingDays -= $holidaysBetween;
+            return $workingDays;
+        } else {
+            return $workingDays = $event->getHours() / $workCalendar->getWorkingHours();
+        }
+    }
+
+    /**
+     * Returns true it it daesn't excess the maximum days
+     */
+    private function checkDoesNotExcessMaximumDaysForType(Event $event, WorkCalendar $workCalendar)
+    {
+        $year = $event->getStartDate()->format('Y');
+        $valid = true;
+        $maxDays = null;
+        if ($event->getType()->getId() === EventType::VACATION) {
+            $maxDays = $workCalendar->getVacationDays();
+        }
+        if ($event->getType()->getId() === EventType::PARTICULAR_BUSSINESS_LEAVE) {
+            $maxDays = $workCalendar->getParticularBusinessLeave();
+        }
+        if ($event->getType()->getId() === EventType::EXCESSIVE_WORKING_HOURS) {
+            $maxDays = $workCalendar->getOvertimeDays();
+        }
+
+        $valid = $this->checkDoesNotExcessMaximumDays($event, $maxDays, $year, $workCalendar);
+        return $valid;
+    }
+
+    private function checkDoesNotExcessMaximumPartionableHours(Event $event, $year, WorkCalendar $workCalendar)
+    {
+        $user = $this->getUser();
+        $totalHours = 0;
+        $eventsThisYear = $this->getDoctrine()->getRepository(Event::class)->findUserEventsCurrentYearAndType($user, $year, $event->getType(), true);
+        foreach ($eventsThisYear as $event) {
+            if (null !== $event->getHours()) {
+                $totalHours += $event->getHours();
+            }
+        }
+        if ($totalHours + $event->getHours() > $workCalendar->getPartitionableHours()) {
+            $this->addFlash('error', $this->translator->trans('message.partitionableHoursExceeded', [
+                'maximumHours' => $workCalendar->getPartitionableHours(),
+                'hours' => $totalHours + $event->getHours(),
+                'year' => $event->getStartDate()->format('Y'),
+            ]));
+            return false;
+        }
+        return true;
+    }
+
+    private function checkDoesNotExcessMaximumDays(Event $event, $maxDays, $year, WorkCalendar $workCalendar)
+    {
+        if (null !== $maxDays) {
+            $user = $this->getUser();
+            $eventsThisYear = $this->getDoctrine()->getRepository(Event::class)->findUserEventsCurrentYearAndType($user, $year, $event->getType());
+            $workingDays = $this->calculateTotalWorkingDays($eventsThisYear, $workCalendar);
+            if ($workingDays + $this->calculateWorkingDays($event, $workCalendar) > $maxDays) {
+                $this->addFlash(
+                    'error',
+                    $this->translator->trans('message.maximum_' . $event->getType()->getId() . '_days_exceeded', [
+                        'days' => $workingDays + $this->calculateWorkingDays($event, $workCalendar) - $maxDays
+                    ])
+                );
+                return false;
+            }
+        }
+        return true;
+    }
+
+    private function renderError($form, $ajax = false)
+    {
+        $template = $ajax ? '_form.html.twig' : 'new.html.twig';
+        return $this->render('event/' . $template, [
+            'event' => $form->getData(),
+            'form' => $form->createView(),
+        ], new Response(null, 422));
     }
 }
